plugins/group_daily_analysis/requirements.txt
nonebot-adapter-onebot
nonebot-plugin-htmlrender
nonebot-plugin-apscheduler
jinja2
httpx
pydantic
